+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+## [Unreleased]
+### Added
+- `getBuildedTransaction` method in to api-caller
+- `needSetSourceAccount` attribute in to payment method(Payment builder)
 
 ## [1.10.3-rc.0] - 2019-10-01
 ### Added